+++ RuntimeEditor/Package.cs
@@ -5,6 +5,6 @@ namespace Hananoki.SharedModule {
   public static class Package {
     public const string name = "SharedModule";
     public const string editorPrefName = "Hananoki.SharedModule";
-    public const string version = "1.7.0";
+    public const string version = "1.7.1";
   }
 }